+++ ceammc/ext/src/data/mod_data.cpp
@@ -44,6 +44,7 @@ void setup_color_scale();
 void setup_data_bimap();
 void setup_data_color();
 void setup_data_copy();
+void setup_global_bimap();
 void setup_local_bimap();
 void setup_rtree_to_list();
 
@@ -92,6 +93,7 @@ void ceammc_data_setup()
     setup_local_mlist();
     setup_local_set();
 
+    setup_global_bimap();
     setup_global_dict();
     setup_global_float();
     setup_global_int();