+++ ceammc/CHANGELOG.md
@@ -27,6 +27,7 @@
   - fx.scramble~ - scramble delayed fragments
   - fx.tremolo~ - tremolo effect from guitarix library
   - gate~ - multichannel signal gate added
+  - global.bimap - global scope bimap container
   - local.bimap - canvas local scope bimap container
   - merge~ - parametrized signal merge
   - metro.choice - randomly (uniformly) metronome time interval choice from specified set